fix(product): address billing merge blockers
