+++ apps/packages/product-surfaces/src/settings/BillingSettingsSurface.tsx
@@ -8,14 +8,17 @@ import {
   BillingSettingsPane,
   type BillingPlanView,
 } from "@proliferate/product-ui/billing/BillingSettingsPane";
-import { SettingsCard } from "@proliferate/product-ui/settings/SettingsCard";
 import { SettingsPageHeader } from "@proliferate/product-ui/settings/SettingsPageHeader";
-import { ExternalLink } from "@proliferate/ui/icons";
-import { Badge } from "@proliferate/ui/primitives/Badge";
-import { Button } from "@proliferate/ui/primitives/Button";
-import { ProgressBar } from "@proliferate/ui/primitives/ProgressBar";
-import { Select } from "@proliferate/ui/primitives/Select";
-import { Switch } from "@proliferate/ui/primitives/Switch";
+import {
+  BillingAutoTopUpCard,
+  BillingPlanCard,
+  BillingPortalCard,
+  type BillingPlanPresentation,
+} from "./BillingManagementCards";
+import {
+  BillingUsageUnitsSection,
+  type BillingUnitBalancePresentation,
+} from "./BillingUsageUnitsSection";
 
 export type BillingCheckoutReturnState = "success" | "cancel" | null;
 
@@ -37,20 +40,6 @@ export interface BillingSettingsSurfaceProps {
   onOpenOrganizationSettings: () => void;
 }
 
-const COMPUTE_TOP_UP_OPTIONS = [
-  { value: "40", label: "40 PCUs - $10.00" },
-  { value: "100", label: "100 PCUs - $25.00" },
-  { value: "200", label: "200 PCUs - $50.00" },
-  { value: "500", label: "500 PCUs - $125.00" },
-];
-
-const LLM_TOP_UP_OPTIONS = [
-  { value: "2500", label: "2,500 LLM credits - $25.00" },
-  { value: "5000", label: "5,000 LLM credits - $50.00" },
-  { value: "10000", label: "10,000 LLM credits - $100.00" },
-  { value: "25000", label: "25,000 LLM credits - $250.00" },
-];
-
 const MOCK_COMPUTE_BALANCE: BillingUnitBalancePresentation = {
   kind: "compute",
   title: "Compute units",
@@ -61,7 +50,6 @@ const MOCK_COMPUTE_BALANCE: BillingUnitBalancePresentation = {
   availablePercent: 33,
   topUpLabel: "Add compute units",
   lowBalanceCopy: "Need more runtime capacity? Add compute units any time.",
-  topUpOptions: COMPUTE_TOP_UP_OPTIONS,
 };
 
 const MOCK_LLM_BALANCE: BillingUnitBalancePresentation = {
@@ -74,7 +62,6 @@ const MOCK_LLM_BALANCE: BillingUnitBalancePresentation = {
   availablePercent: 38,
   topUpLabel: "Add LLM credits",
   lowBalanceCopy: "Need more model usage? Add LLM credits any time.",
-  topUpOptions: LLM_TOP_UP_OPTIONS,
 };
 
 export function BillingSettingsSurface({
@@ -95,8 +82,6 @@ export function BillingSettingsSurface({
   const comparisonActions = useCloudBillingActions(comparisonOwner, billingReturnOptions);
   const [comparisonActionError, setComparisonActionError] = useState<string | null>(null);
   const [planManagementOpen, setPlanManagementOpen] = useState(false);
-  const [computeTopUpAmount, setComputeTopUpAmount] = useState("100");
-  const [llmTopUpAmount, setLlmTopUpAmount] = useState("5000");
 
   useEffect(() => {
     if (checkoutReturnState !== "success") {
@@ -118,14 +103,12 @@ export function BillingSettingsSurface({
     setPlanManagementOpen(true);
   }
 
-  async function openComparisonBillingAction(action: "checkout" | "portal" | "refill") {
+  async function openComparisonBillingAction(action: "checkout" | "portal") {
     setComparisonActionError(null);
     try {
       const response = action === "portal"
         ? await comparisonActions.createBillingPortal()
-        : action === "refill"
-          ? await comparisonActions.createRefillCheckout()
-          : await comparisonActions.createCloudCheckout();
+        : await comparisonActions.createCloudCheckout();
       await onOpenUrl(response.url);
     } catch (error) {
       setComparisonActionError(
@@ -234,32 +217,16 @@ export function BillingSettingsSurface({
           onManage={openPlanManagement}
         />
 
-        <BillingUnitBalancesCard
+        <BillingUsageUnitsSection
           unitBalances={unitBalances}
-          topUpAmounts={{
-            compute: computeTopUpAmount,
-            llm: llmTopUpAmount,
-          }}
-          addCreditsLoading={comparisonActions.creatingRefillCheckout}
-          addCreditsDisabled={billingActionDisabled}
-          onTopUpAmountChange={(kind, value) => {
-            if (kind === "compute") {
-              setComputeTopUpAmount(value);
-              return;
-            }
-            setLlmTopUpAmount(value);
-          }}
-          onAddCredits={(_kind) => {
-            void openComparisonBillingAction("refill");
-          }}
+          addCreditsLoading={false}
+          addCreditsDisabled
         />
 
         <BillingAutoTopUpCard
           enabled={topUpEnabled}
           disabled={billingActionDisabled || comparisonActions.updatingOverage}
           saving={comparisonActions.updatingOverage}
-          computeTopUpLabel={labelForTopUpOption(COMPUTE_TOP_UP_OPTIONS, computeTopUpAmount)}
-          llmTopUpLabel={labelForTopUpOption(LLM_TOP_UP_OPTIONS, llmTopUpAmount)}
           onEnabledChange={(value) => {
             void updateComparisonTopUp(value);
           }}
@@ -277,297 +244,7 @@ export function BillingSettingsSurface({
   );
 }
 
-function BillingPlanCard({
-  plan,
-  organization,
-  organizationLoading,
-  loading,
-  actionError,
-  onManage,
-}: {
-  plan: BillingPlanPresentation;
-  organization: BillingSettingsOrganization | null;
-  organizationLoading: boolean;
-  loading: boolean;
-  actionError: string | null;
-  onManage: () => void;
-}) {
-  return (
-    <SettingsCard>
-      <div className="space-y-4 p-5">
-        <h2 className="text-lg font-semibold text-foreground">Plan</h2>
-        <div className="flex flex-col gap-3 sm:flex-row sm:items-start sm:justify-between">
-          <div className="min-w-0 space-y-1">
-            <div className="flex flex-wrap items-center gap-1.5">
-              <span className="text-base font-medium text-foreground">{plan.name}</span>
-              <span className="text-base font-medium text-muted-foreground">· {plan.price}</span>
-              <Badge tone={plan.badgeTone}>{plan.badge}</Badge>
-            </div>
-            <p className="text-sm leading-5 text-muted-foreground">
-              {organizationLoading
-                ? "Billing details for this organization."
-                : organization
-                  ? `Billing for ${organization.name}.`
-                  : "Shared billing and credits for this workspace."}
-            </p>
-            {actionError ? <p className="text-sm text-destructive">{actionError}</p> : null}
-          </div>
-          <Button
-            type="button"
-            variant="primary"
-            loading={loading}
-            disabled={organizationLoading}
-            onClick={onManage}
-            className="w-full sm:w-auto"
-          >
-            Manage
-          </Button>
-        </div>
-      </div>
-    </SettingsCard>
-  );
-}
-
-function BillingUnitBalancesCard({
-  unitBalances,
-  topUpAmounts,
-  addCreditsLoading,
-  addCreditsDisabled,
-  onTopUpAmountChange,
-  onAddCredits,
-}: {
-  unitBalances: BillingUnitBalancePresentation[];
-  topUpAmounts: Record<BillingUnitKind, string>;
-  addCreditsLoading: boolean;
-  addCreditsDisabled: boolean;
-  onTopUpAmountChange: (kind: BillingUnitKind, value: string) => void;
-  onAddCredits: (kind: BillingUnitKind) => void;
-}) {
-  return (
-    <SettingsCard>
-      <div className="space-y-6 p-5">
-        <div className="space-y-1.5">
-          <h2 className="text-lg font-semibold text-foreground">Usage units</h2>
-          <p className="max-w-2xl text-sm leading-6 text-muted-foreground">
-            Compute units and LLM credits are tracked, budgeted, and topped up separately.
-          </p>
-        </div>
-
-        <div className="grid gap-4">
-          {unitBalances.map((unitBalance) => (
-            <BillingUnitPoolCard
-              key={unitBalance.kind}
-              unitBalance={unitBalance}
-              topUpAmount={topUpAmounts[unitBalance.kind]}
-              addCreditsLoading={addCreditsLoading}
-              addCreditsDisabled={addCreditsDisabled}
-              onTopUpAmountChange={(value) => onTopUpAmountChange(unitBalance.kind, value)}
-              onAddCredits={() => onAddCredits(unitBalance.kind)}
-            />
-          ))}
-        </div>
-      </div>
-    </SettingsCard>
-  );
-}
-
-function BillingUnitPoolCard({
-  unitBalance,
-  topUpAmount,
-  addCreditsLoading,
-  addCreditsDisabled,
-  onTopUpAmountChange,
-  onAddCredits,
-}: {
-  unitBalance: BillingUnitBalancePresentation;
-  topUpAmount: string;
-  addCreditsLoading: boolean;
-  addCreditsDisabled: boolean;
-  onTopUpAmountChange: (value: string) => void;
-  onAddCredits: () => void;
-}) {
-  const selectId = `billing-${unitBalance.kind}-top-up-amount`;
-
-  return (
-    <section className="overflow-hidden rounded-lg border border-border-light bg-foreground/[0.02]">
-      <div className="space-y-5 p-4">
-        <div className="space-y-1">
-          <h3 className="text-base font-semibold text-foreground">{unitBalance.title}</h3>
-          <p className="text-sm leading-5 text-muted-foreground">{unitBalance.description}</p>
-        </div>
-
-        <div className="grid gap-3 sm:grid-cols-3">
-          <BillingMetric label="Total purchased" value={unitBalance.purchased} />
-          <BillingMetric label="Available" value={unitBalance.available} />
-          <BillingMetric label="Used" value={unitBalance.used} />
-        </div>
-
-        <ProgressBar
-          value={unitBalance.availablePercent ?? 0}
-          className="h-4 overflow-hidden rounded-full border border-border-light bg-foreground/5 p-0.5"
-          indicatorClassName="h-full rounded-full bg-primary/70"
-          aria-label={`${unitBalance.title} available`}
-        />
-      </div>
-
-      <div className="border-t border-border-light bg-background/40 p-4">
-        <div className="flex flex-col gap-1 sm:flex-row sm:items-end sm:justify-between">
-          <div className="min-w-0">
-            <h4 className="text-sm font-medium text-foreground">
-              Top up {unitBalance.title.toLowerCase()}
-            </h4>
-            <p className="mt-1 text-sm leading-5 text-muted-foreground">
-              {unitBalance.lowBalanceCopy}
-            </p>
-          </div>
-        </div>
-        <div className="mt-3 grid gap-2 sm:grid-cols-[minmax(0,1fr)_auto]">
-          <div className="min-w-0 flex-1">
-            <label className="sr-only" htmlFor={selectId}>
-              {unitBalance.title} top-up amount
-            </label>
-            <Select
-              id={selectId}
-              value={topUpAmount}
-              onChange={(event) => onTopUpAmountChange(event.currentTarget.value)}
-              disabled={addCreditsDisabled}
-            >
-              {unitBalance.topUpOptions.map((option) => (
-                <option key={option.value} value={option.value}>{option.label}</option>
-              ))}
-            </Select>
-          </div>
-          <Button
-            type="button"
-            variant="secondary"
-            loading={addCreditsLoading}
-            disabled={addCreditsDisabled}
-            onClick={onAddCredits}
-            className="w-full sm:w-auto"
-          >
-            {unitBalance.topUpLabel}
-          </Button>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function BillingAutoTopUpCard({
-  enabled,
-  disabled,
-  saving,
-  computeTopUpLabel,
-  llmTopUpLabel,
-  onEnabledChange,
-}: {
-  enabled: boolean;
-  disabled: boolean;
-  saving: boolean;
-  computeTopUpLabel: string;
-  llmTopUpLabel: string;
-  onEnabledChange: (value: boolean) => void;
-}) {
-  return (
-    <SettingsCard>
-      <div className="space-y-6 p-5">
-        <div className="space-y-1.5">
-          <h2 className="text-lg font-semibold text-foreground">Auto top-up</h2>
-          <p className="max-w-2xl text-sm leading-6 text-muted-foreground">
-            Automatically replenish compute units and LLM credits from their own thresholds.
-          </p>
-        </div>
-
-        <div className="flex items-center gap-3">
-          <Switch
-            checked={enabled}
-            onChange={onEnabledChange}
-            disabled={disabled}
-            aria-label="Auto top-up"
-          />
-          <span className="text-sm font-medium text-foreground">Enable auto top-up</span>
-          {saving ? <span className="text-xs text-muted-foreground">Saving...</span> : null}
-        </div>
-
-        <div className="grid gap-3 sm:grid-cols-2">
-          <AutoTopUpSummary
-            label="Compute auto top-up"
-            description={`When compute units fall below 40 PCUs, add ${computeTopUpLabel}.`}
-          />
-          <AutoTopUpSummary
-            label="LLM credit auto top-up"
-            description={`When LLM credits fall below 2,000, add ${llmTopUpLabel}.`}
-          />
-        </div>
-
-        <p className="text-sm text-muted-foreground">Last auto top-up: not yet run.</p>
-      </div>
-    </SettingsCard>
-  );
-}
-
-function AutoTopUpSummary({ label, description }: { label: string; description: string }) {
-  return (
-    <div className="rounded-lg border border-border-light bg-foreground/[0.02] p-3">
-      <div className="text-sm font-medium text-foreground">{label}</div>
-      <div className="mt-1 text-sm leading-5 text-muted-foreground">{description}</div>
-    </div>
-  );
-}
-
-function BillingPortalCard({
-  loading,
-  disabled,
-  onOpenPortal,
-}: {
-  loading: boolean;
-  disabled: boolean;
-  onOpenPortal: () => void;
-}) {
-  return (
-    <SettingsCard>
-      <div className="flex flex-col gap-4 p-5 sm:flex-row sm:items-center sm:justify-between">
-        <div className="min-w-0 space-y-1.5">
-          <h2 className="text-lg font-semibold text-foreground">Billing</h2>
-          <p className="max-w-2xl text-sm leading-6 text-muted-foreground">
-            View invoices, update payment methods, and manage cancellation in Stripe.
-          </p>
-        </div>
-        <Button
-          type="button"
-          variant="primary"
-          loading={loading}
-          disabled={disabled}
-          onClick={onOpenPortal}
-          className="w-full sm:w-auto"
-        >
-          Access billing portal
-          <ExternalLink className="size-3.5" />
-        </Button>
-      </div>
-    </SettingsCard>
-  );
-}
-
-function BillingMetric({ label, value }: { label: string; value: string }) {
-  return (
-    <div className="min-w-0">
-      <p className="mb-1 text-sm text-muted-foreground">{label}</p>
-      <p className="break-words text-xl font-semibold leading-tight text-foreground sm:text-2xl">
-        {value}
-      </p>
-    </div>
-  );
-}
-
 type BillingPlanViewKey = "free" | "core" | "enterprise";
-
-interface BillingPlanPresentation {
-  name: string;
-  price: string;
-  badge: string;
-  badgeTone: "neutral" | "success" | "info" | "warning" | "destructive";
-}
 
 function planKeyForBilling(plan: BillingPlanView | null | undefined): BillingPlanViewKey | null {
   if (!plan) {
@@ -613,26 +290,6 @@ function planSummary(
     badge: "Active",
     badgeTone: "neutral",
   };
-}
-
-type BillingUnitKind = "compute" | "llm";
-
-interface BillingTopUpOption {
-  value: string;
-  label: string;
-}
-
-interface BillingUnitBalancePresentation {
-  kind: BillingUnitKind;
-  title: string;
-  description: string;
-  purchased: string;
-  available: string;
-  used: string;
-  availablePercent: number | null;
-  topUpLabel: string;
-  lowBalanceCopy: string;
-  topUpOptions: BillingTopUpOption[];
 }
 
 function billingUnitBalances(
@@ -692,8 +349,4 @@ function formatCredits(value: number | null | undefined): string {
     maximumFractionDigits: Number.isInteger(rounded) ? 0 : 1,
   });
   return `${formatted} ${rounded === 1 ? "PCU" : "PCUs"}`;
-}
-
-function labelForTopUpOption(options: BillingTopUpOption[], value: string): string {
-  return options.find((option) => option.value === value)?.label ?? options[0]?.label ?? value;
 }
